refactor: global 폴더로 전역적인 코드를 관리하도록 변경
