+++ src/main/java/com/ziine/admin/auth/presentation/AdminAuthenticationInterceptor.java
@@ -1,9 +1,9 @@
 package com.ziine.admin.auth.presentation;
 
 import com.ziine.admin.auth.application.AdminContextHolder;
-import com.ziine.admin.auth.application.JwtService;
 import com.ziine.admin.auth.application.exception.AdminUnauthorizedException;
 import com.ziine.admin.auth.domain.Admin;
+import com.ziine.global.jwt.application.JwtService;
 import io.jsonwebtoken.Claims;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
