+++ src/main/java/com/ziine/global/jpa/config/JpaAuditingConfig.java
@@ -1,4 +1,4 @@
-package com.ziine.common.config;
+package com.ziine.global.jpa.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
