+++ src/main/java/com/ziine/global/jpa/domain/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.ziine.common.entity;
+package com.ziine.global.jpa.domain.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
