+++ src/main/java/com/ziine/global/swagger/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.ziine.external.swagger.config;
+package com.ziine.global.swagger.config;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
